move commands to command package
The internal/controller/commands package is unnecessary and makes things
more complicated than they need to be.

Commands were moved into internal/command. Since this is an internal
package this is not a breaking change.
