+++ internal/statedrift/detector.go
@@ -5,7 +5,6 @@ package statedrift
 
 import (
 	"github.com/martinohmann/rfoutlet/internal/command"
-	"github.com/martinohmann/rfoutlet/internal/controller/commands"
 	"github.com/martinohmann/rfoutlet/internal/outlet"
 	"github.com/martinohmann/rfoutlet/pkg/gpio"
 	"github.com/sirupsen/logrus"
@@ -50,13 +49,13 @@ func (d *Detector) Run(stopCh <-chan struct{}) {
 			for _, o := range d.Registry.GetOutlets() {
 				if result.Code == o.CodeOn && o.GetState() != outlet.StateOn {
 					found = true
-					d.CommandQueue <- commands.StateCorrectionCommand{
+					d.CommandQueue <- command.StateCorrectionCommand{
 						Outlet:       o,
 						DesiredState: outlet.StateOn,
 					}
 				} else if result.Code == o.CodeOff && o.GetState() != outlet.StateOff {
 					found = true
-					d.CommandQueue <- commands.StateCorrectionCommand{
+					d.CommandQueue <- command.StateCorrectionCommand{
 						Outlet:       o,
 						DesiredState: outlet.StateOff,
 					}
