+++ internal/statedrift/detector_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	"github.com/martinohmann/rfoutlet/internal/command"
-	"github.com/martinohmann/rfoutlet/internal/controller/commands"
 	"github.com/martinohmann/rfoutlet/internal/outlet"
 	"github.com/martinohmann/rfoutlet/pkg/gpio"
 	"github.com/stretchr/testify/assert"
@@ -49,8 +48,8 @@ func TestDetector(t *testing.T) {
 	}()
 
 	expected := []command.Command{
-		commands.StateCorrectionCommand{Outlet: o1, DesiredState: outlet.StateOn},
-		commands.StateCorrectionCommand{Outlet: o2, DesiredState: outlet.StateOff},
+		command.StateCorrectionCommand{Outlet: o1, DesiredState: outlet.StateOn},
+		command.StateCorrectionCommand{Outlet: o2, DesiredState: outlet.StateOff},
 	}
 
 	received := make([]command.Command, 0)
