+++ internal/websocket/client_test.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/martinohmann/rfoutlet/internal/command"
-	"github.com/martinohmann/rfoutlet/internal/controller/commands"
 	"github.com/posener/wstest"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -29,14 +28,14 @@ func TestClient_listenRead(t *testing.T) {
 					"id":     "foo",
 				},
 			},
-			expectedCmdType: &commands.OutletCommand{},
+			expectedCmdType: &command.OutletCommand{},
 		},
 		{
 			name: "status action",
 			data: map[string]interface{}{
 				"type": "status",
 			},
-			expectedCmdType: &commands.StatusCommand{},
+			expectedCmdType: &command.StatusCommand{},
 		},
 		{
 			name: "unknown command",
